# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Goblin L00t API - Twitch bot + giveaway dashboard
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /giveaway-presets:
    get:
      operationId: listGiveawayPresets
      tags: [giveaway]
      summary: List the caller's saved giveaway presets
      responses:
        "200":
          description: Preset list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GiveawayPreset"
        "401":
          description: Unauthorized
    post:
      operationId: createGiveawayPreset
      tags: [giveaway]
      summary: Save a new giveaway preset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGiveawayPreset"
      responses:
        "201":
          description: Preset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiveawayPreset"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
  /giveaway-presets/{id}:
    delete:
      operationId: deleteGiveawayPreset
      tags: [giveaway]
      summary: Delete a saved preset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /giveaway-presets/{id}/launch:
    post:
      operationId: launchGiveawayPreset
      tags: [giveaway]
      summary: Create a new pending giveaway from a preset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Giveaway created from preset
          content:
            application/json:
              schema:
                type: object
                properties:
                  giveawayId:
                    type: integer
                  status:
                    type: string
                required: [giveawayId, status]
        "401":
          description: Unauthorized
        "404":
          description: Not found
...
  /chat-users:
    get:
      operationId: listChatUsers
      tags: [chat-users]
      summary: List all chat users in the streamer's channel with coin balances and inventory
      responses:
        "200":
          description: List of chat users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatUser"
        "401":
          description: Unauthorized
  /chat-users/{username}/coins:
    post:
      operationId: adjustChatUserCoins
      tags: [chat-users]
      summary: Adjust a chat user's coin balance (positive or negative delta)
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustCoinsRequest"
      responses:
        "200":
          description: Adjustment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdjustCoinsResponse"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
...
components:
...
  schemas:
...
    ChatUserInventoryItem:
      type: object
      properties:
        id:
          type: integer
        slot:
          type: integer
        item:
          type: string
        rarity:
          type: string
          enum: [common, uncommon, rare, epic, legendary]
        kind:
          type: string
      required:
        - id
        - slot
        - item
        - rarity
        - kind
    ChatUser:
      type: object
      properties:
        username:
          type: string
        coins:
          type: integer
        inventoryCount:
          type: integer
        inventory:
          type: array
          items:
            $ref: "#/components/schemas/ChatUserInventoryItem"
      required:
        - username
        - coins
        - inventoryCount
        - inventory
    AdjustCoinsRequest:
      type: object
      properties:
        delta:
          type: integer
        reason:
          type: string
      required:
        - delta
    AdjustCoinsResponse:
      type: object
      properties:
        ok:
          type: boolean
        username:
          type: string
        balance:
          type: integer
      required:
        - ok
        - username
        - balance
...
    GiveawayPreset:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        prize:
          type: string
        prizeKind:
          type: string
          enum: [cs2, bot_item, bot_coins]
        prizeBotCoins:
          type: ["integer", "null"]
        prizeBotRarity:
          type: ["string", "null"]
        keyword:
          type: string
        requireFollower:
          type: boolean
        subscriberOnly:
          type: boolean
        minSubTier:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - prize
        - prizeKind
        - keyword
        - requireFollower
        - subscriberOnly
        - createdAt
    CreateGiveawayPreset:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        prize:
          type: string
        prizeKind:
          type: string
          enum: [cs2, bot_item, bot_coins]
        prizeBotCoins:
          type: integer
        prizeBotRarity:
          type: string
          enum: [common, uncommon, rare, epic, legendary]
        keyword:
          type: string
        requireFollower:
          type: boolean
        subscriberOnly:
          type: boolean
        minSubTier:
          type: string
          enum: ["1000", "2000", "3000"]
      required:
        - title
        - prize