# lib/api-zod/src/generated/types/createGiveawayPreset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Goblin L00t API - Twitch bot + giveaway dashboard
 * OpenAPI spec version: 0.1.0
 */
import type { CreateGiveawayPresetMinSubTier } from "./createGiveawayPresetMinSubTier";
import type { CreateGiveawayPresetPrizeBotRarity } from "./createGiveawayPresetPrizeBotRarity";
import type { CreateGiveawayPresetPrizeKind } from "./createGiveawayPresetPrizeKind";

export interface CreateGiveawayPreset {
  title: string;
  description?: string;
  prize: string;
  prizeKind?: CreateGiveawayPresetPrizeKind;
  prizeBotCoins?: number;
  prizeBotRarity?: CreateGiveawayPresetPrizeBotRarity;
  keyword?: string;
  requireFollower?: boolean;
  subscriberOnly?: boolean;
  minSubTier?: CreateGiveawayPresetMinSubTier;
}